Remove helper function for header/source switching
The function no longer conforms the way project files
are structured. It is also not trivial to write a function
which would work 100% as the 'project name' folder in
'include' folder can not easily be retrieved.

Therefore, the new workflow for header/source switching
is to use the fuzzy file finder or the languageserver
'go to definition' functionality.